melo_tts_api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
torch
torchaudio
